requirements.txt
Django==2.2.6
django-cors-headers==3.2.0
django-widget-tweaks==1.4.5
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
PyJWT==1.7.1
pytz==2019.2
sqlparse==0.3.0
